refactor(sso): code-review + simplify cleanups
- Exclude amr/acr/sid (auth-method/session claims) from claim suggestions so
  they aren't offered as role mappings; drop empty-string claim values.
- Autocomplete: build the cmdk filter value without trailing spaces.
- ClaimMappingEditor: seed the disclosure open state with hasSuggestions to
  avoid a closed-then-open flash.
- Move JsonValue to lib/shared/json so shared/client code stops importing it
  from lib/server (re-exported from audit/log for existing consumers).
- Correct the stale legacy ssoOidc.attributeMapping JSDoc (no-match returns
  null now). Tidy a test's import/setup order and use a generic IdP fixture.
