+++ apps/web/src/components/admin/settings/security/identity-providers/__tests__/provider-editor.test.tsx
@@ -11,6 +11,10 @@
  */
 import { describe, it, expect, beforeAll, beforeEach, vi } from 'vitest'
 import { render, screen, fireEvent, waitFor } from '@testing-library/react'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import type { IdentityProviderId } from '@quackback/ids'
+import type { IdentityProvider } from '@/lib/server/domains/settings/identity-providers.service'
+import { ProviderEditor } from '../provider-editor'
 
 beforeAll(() => {
   Element.prototype.scrollIntoView = vi.fn()
@@ -18,10 +22,6 @@ beforeAll(() => {
   Element.prototype.setPointerCapture = vi.fn()
   Element.prototype.releasePointerCapture = vi.fn()
 })
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
-import type { IdentityProviderId } from '@quackback/ids'
-import type { IdentityProvider } from '@/lib/server/domains/settings/identity-providers.service'
-import { ProviderEditor } from '../provider-editor'
 
 const { upsertSpy } = vi.hoisted(() => ({
   upsertSpy: vi.fn(
